mcnode refactor: split out node specific code, some improvements in publish/retrieve
